requirements.txt
pdfplumber>=0.11.0
python-dateutil>=2.9.0
rapidfuzz>=3.9.0
pydantic>=2.7.0